LPS-81743 Create a new object that contains the stores instances during the data migration
